Example works in JBoss AS 6
Using a hudson build from today I was able to get this to work!

Had to make some changes for seam xml, but not really anything major.  Also changed the services to lookup by id instead of name.

Added jboss as specific info in the pom and also for resources.