Fixing handlers to act the same as observers wrt qualifiers.
I had a wrong assumption about observers because I was going off of the Weld Docs, which were wrong. I've spoken to Pete, fixed the Weld Docs and now am fixing Catch.

A handler (or observer for that matter) must exactly match the qualifiers (no extra, nor lacking) to be invoked.