Fixing up tests to run in IDEA
I don't know if this had anything to do with it or not, or if it was
just me setting up IDEA correctly, but I can now run the tests for the
various test suites.

I think besides setting up IDEA correctly was getting the unpacking done
at a phase I can tell IDEA to execute. I sincerely hope eclipse and m2e
can be setup similarly.